refactor(hy_vla): consolidate policy modeling code
